requirements.txt
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
openpyxl>=3.0.0
scipy>=1.7.0
scikit-learn>=1.0.0
python-docx>=0.8.11 